// app/api/operator/agent/types.ts
export type FunctionOutput = {
  type: "function_call_output";
  call_id: string;
  output: string;
};

export type ComputerCallOutput = {
  type: "computer_call_output";
  call_id: string;
  output: { type: "input_image"; image_url: string };
  acknowledged_safety_checks: SafetyCheck[];
  current_url?: string;
};
// ...
export type Item = Message | FunctionToolCall | ComputerToolCall | Reasoning;

export type Message = {
  id: string;
  type: "message";
  role: "user" | "assistant" | "developer" | "system";
  content: Content[];
};

export type Reasoning = {
  id: string;
  type: "reasoning";
  content: [];
};

export type FunctionToolCall = {
  type: "function_call";
  id: string;
  call_id: string;
  name: string;
  arguments: string;
  output: Content[] | null;
};

export type ComputerAction =
  | Click
  | DoubleClick
  | Drag
  | Screenshot
  | KeyPress
  | Move
  | Scroll
  | Type
  | Wait;

export type ComputerToolCall = {
  type: "computer_call";
  id: string;
  call_id: string;
  action: ComputerAction;
  pending_safety_checks: SafetyCheck[];
};
// ...
export type SafetyCheck = {
  id: string;
  code: string;
  message: string;
};

export type InputContent = InputText | InputImage | InputFile;

export type OutputContent = OutputText | Refusal;

export type Content = InputContent | OutputContent | Reasoning;